+++ R/discretize_xgb.R
@@ -30,8 +30,8 @@
 #'  each variable.
 #' @param id A character string that is unique to this step to identify it.
 #' @param skip A logical. Should the step be skipped when the
-#'  recipe is baked by [recipes::bake.recipe()]? While all operations are baked
-#'  when [recipes::prep.recipe()] is run, some operations may not be able to be
+#'  recipe is baked by [recipes::bake()]? While all operations are baked
+#'  when [recipes::prep()] is run, some operations may not be able to be
 #'  conducted on new data (e.g. processing the outcome variable(s)).
 #'  Care should be taken when using `skip = TRUE` as it may affect
 #'  the computations for subsequent operations
@@ -61,6 +61,11 @@
 #'
 #' Note that the original data will be replaced with the new bins.
 #' 
+#' # Tidying
+#'
+#' When you [`tidy()`][tidy.recipe()] this step, a tibble with columns
+#' `terms` (the columns that is selected), `values` is returned.
+#' 
 #' @template case-weights-supervised
 #'
 #' @examples
@@ -85,7 +90,7 @@
 #'   bake(xgb_rec, credit_data_te, Price)
 #' }
 #' @seealso [embed::step_discretize_cart()], [recipes::recipe()],
-#' [recipes::prep.recipe()], [recipes::bake.recipe()]
+#' [recipes::prep()], [recipes::bake()]
 
 step_discretize_xgb <-
   function(recipe,
@@ -466,7 +471,7 @@ bake.step_discretize_xgb <- function(object, new_data, ...) {
       new_data <- binned_data
     }
   }
-  tibble::as_tibble(new_data)
+  new_data
 }
 
 #' @export
@@ -477,7 +482,7 @@ print.step_discretize_xgb <- function(x, width = max(20, options()$width - 30), 
   invisible(x)
 }
 
-#' @rdname step_discretize_xgb
+#' @rdname tidy.recipe
 #' @param x A `step_discretize_xgb` object.
 #' @export
 tidy.step_discretize_xgb <- function(x, ...) {