+++ R/pca_sparse.R
@@ -22,7 +22,7 @@
 #'  output. Defaults to `FALSE`.
 #' @param options A list of options to the default method for [irlba::ssvd()].
 #' @param res The rotation matrix once this
-#'  preprocessing step has be trained by [prep.recipe()].
+#'  preprocessing step has be trained by [prep()].
 #' @param prefix A character string that will be the prefix to the resulting
 #'  new variables. See notes below.
 #' @return An updated version of `recipe` with the new step added to the
@@ -48,6 +48,12 @@
 #'  If `num_comp = 101`, the names would be `PC001` -
 #'  `PC101`.
 #'  
+#' # Tidying
+#' 
+#' When you [`tidy()`][tidy.recipe()] this step, a tibble with columns
+#' `terms` (the selectors or variables selected), `value` and `component` is
+#' returned.
+#' 
 #' @template case-weights-not-supported
 #'
 #' @seealso [step_pca_sparse_bayes()]
@@ -194,7 +200,7 @@ bake.step_pca_sparse <- function(object, new_data, ...) {
       new_data <- new_data[, !(colnames(new_data) %in% pca_vars), drop = FALSE]
     }
   }
-  as_tibble(new_data)
+  new_data
 }
 
 #' @export
@@ -228,22 +234,7 @@ pca_coefs <- function(x) {
   res
 }
 
-#' @rdname step_pca_sparse
-#' @param x A `step_pca_sparse` object.
-#' @export
-tidy.step_pca_sparse <- function(x, ...) {
-  if (!is_trained(x)) {
-    term_names <- sel2char(x$terms)
-    res <- tibble(terms = term_names, value = na_dbl, component = na_chr)
-  } else {
-    res <- pca_coefs(x)
-  }
-  res$id <- x$id
-  res
-}
-
-
-#' @rdname step_pca_sparse
+#' @rdname tidy.recipe
 #' @param x A `step_pca_sparse` object.
 #' @export
 tidy.step_pca_sparse <- function(x, ...) {