+++ NEWS.md
@@ -1,6 +1,18 @@
 # embed (development version)
 
+* `step_collapse_cart()` can pool a predictor's factor levels using a tree-based method. 
+
+# embed 0.2.0
+
 * `step_embed()` now correctly defaults to have a random id with the word "embed". (#102)
+
+* `step_feature_hash()` is soft deprecated in embed in favor of `step_dummy_hash()` in textrecipes. (#95)
+
+* Steps now have a dedicated subsection detailing what happens when `tidy()` is applied. (#105)
+
+* Reorganize documentation for all recipe step `tidy` methods (#115).
+
+* Fixed a bug where `woe_table()` and `step_woe()` didn't respect the factor levels of the outcome. (109)
 
 # embed 0.1.5